+++ BE/src/main/java/com/newsainturtle/shadowmate/social/dto/response/SearchSocialPlannerResponse.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.social.dto;
+package com.newsainturtle.shadowmate.social.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -8,12 +8,8 @@ import java.util.List;
 @Getter
 @Builder
 public class SearchSocialPlannerResponse {
-
     private long pageNumber;
-
     private long totalPage;
-
     private String sort;
-
     private List<SearchSocialResponse> socialList;
 }
