+++ BE/src/main/java/com/newsainturtle/shadowmate/social/dto/response/SearchSocialResponse.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.social.dto;
+package com.newsainturtle.shadowmate.social.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,19 +6,12 @@ import lombok.Getter;
 @Getter
 @Builder
 public class SearchSocialResponse {
-
     private long socialId;
-
     private String socialImage;
-
     private String dailyPlannerDay;
-
     private long userId;
-
     private String statusMessage;
-
     private String nickname;
-
     private String profileImage;
 
 }
