+++ BE/src/test/java/com/newsainturtle/shadowmate/yn/social/SocialControllerTest.java
@@ -6,8 +6,8 @@ import com.newsainturtle.shadowmate.auth.exception.AuthException;
 import com.newsainturtle.shadowmate.auth.service.AuthServiceImpl;
 import com.newsainturtle.shadowmate.common.GlobalExceptionHandler;
 import com.newsainturtle.shadowmate.social.controller.SocialController;
-import com.newsainturtle.shadowmate.social.dto.SearchSocialPlannerResponse;
-import com.newsainturtle.shadowmate.social.dto.SearchSocialResponse;
+import com.newsainturtle.shadowmate.social.dto.response.SearchSocialPlannerResponse;
+import com.newsainturtle.shadowmate.social.dto.response.SearchSocialResponse;
 import com.newsainturtle.shadowmate.social.entity.Social;
 import com.newsainturtle.shadowmate.social.exception.SocialErrorResult;
 import com.newsainturtle.shadowmate.social.exception.SocialException;
